requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.0
pymysql==1.1.0
requests==2.31.0
pytest==7.4.3